pyproject.toml
[project]
name = "xfinder"
version = "0.1.0"
description = "命令行本地文件搜索工具"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pyyaml",
    "click",
    "colorama",
    "tqdm",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.2",
    "flet>=0.21.0",
]

[project.scripts]
xfinder = "xfinder.main:cli"
xfinder-search = "xfinder.main:search"


